move options to hpp
