+++ src/gennylib/include/gennylib/ExecutionStrategy.hpp
@@ -20,9 +20,68 @@
 
 #include <loki/ScopeGuard.h>
 
-#include <gennylib/config/ExecutionStrategyOptions.hpp>
-
 #include <metrics/metrics.hpp>
+
+namespace genny::config {
+
+/**
+ * Configuration for a `genny::ExecutionStrategy`.
+ */
+struct ExecutionStrategyOptions {
+    /** Default values for each key */
+    struct Defaults {
+        static constexpr auto kMaxRetries = size_t{0};
+        static constexpr auto kThrowOnFailure = false;
+    };
+
+    /** YAML keys to use */
+    struct Keys {
+        static constexpr auto kMaxRetries = "Retries";
+        static constexpr auto kThrowOnFailure = "ThrowOnFailure";
+    };
+
+    size_t maxRetries = Defaults::kMaxRetries;
+    bool throwOnFailure = Defaults::kThrowOnFailure;
+};
+
+}  // namespace genny::config
+
+namespace YAML {
+
+/**
+ * Convert to/from `genny::config::ExecutionStrategyOptions` and YAML
+ */
+template <>
+struct convert<genny::config::ExecutionStrategyOptions> {
+    using Config = genny::config::ExecutionStrategyOptions;
+    using Defaults = typename Config::Defaults;
+    using Keys = typename Config::Keys;
+
+    static Node encode(const Config& rhs) {
+        Node node;
+
+        node[Keys::kMaxRetries] = rhs.maxRetries;
+        node[Keys::kThrowOnFailure] = rhs.throwOnFailure;
+
+        return node;
+    }
+
+    static bool decode(const Node& node, Config& rhs) {
+        if (!node.IsMap()) {
+            return false;
+        }
+
+        genny::decodeNodeInto(rhs.maxRetries, node[Keys::kMaxRetries], Defaults::kMaxRetries);
+        genny::decodeNodeInto(
+                rhs.throwOnFailure, node[Keys::kThrowOnFailure], Defaults::kThrowOnFailure);
+
+        return true;
+    }
+};
+
+}  // namespace YAML
+
+
 
 namespace genny {
